Move invalidation function to LayoutHelper.
